+++ FDB.Apollo.IPT/FDB.Apollo.IPT/Models.EF/IptColorC.cs
@@ -5,17 +5,17 @@ namespace FDB.Apollo.IPT.Service.Models.EF
 {
     public partial class IptColorC
     {
-        public decimal Id { get; set; }
-        public decimal RevNbr { get; set; }
+        public long Id { get; set; }
+        public int RevNbr { get; set; }
         public string Description { get; set; } = null!;
-        public decimal BasicColorId { get; set; }
+        public long BasicColorId { get; set; }
         public string? Abbreviation { get; set; }
         public char DoNotUseInd { get; set; }
         public char ChangeType { get; set; }
-        public decimal ChangeUserId { get; set; }
+        public long ChangeUserId { get; set; }
         public DateTime ChangeTimestamp { get; set; }
-        public decimal ConceptRevNbr { get; set; }
-        public decimal? DcrNbr { get; set; }
+        public int ConceptRevNbr { get; set; }
+        public int? DcrNbr { get; set; }
         public string? LegacyChangeUser { get; set; }
     }
 }
