+++ FDB.Apollo.IPT/FDB.Apollo.IPT/Models.EF/IptColorH.cs
@@ -5,12 +5,12 @@ namespace FDB.Apollo.IPT.Service.Models.EF
 {
     public partial class IptColorH
     {
-        public decimal Id { get; set; }
-        public decimal RevNbr { get; set; }
-        public decimal ChangeUserId { get; set; }
+        public long Id { get; set; }
+        public int RevNbr { get; set; }
+        public long ChangeUserId { get; set; }
         public DateTime ChangeTimestamp { get; set; }
         public char ChangeType { get; set; }
-        public decimal? DcrNbr { get; set; }
+        public int? DcrNbr { get; set; }
         public string? LegacyChangeUser { get; set; }
     }
 }
