+++ FDB.Apollo.IPT/FDB.Apollo.IPT/Models.EF/IptColorP.cs
@@ -5,9 +5,9 @@ namespace FDB.Apollo.IPT.Service.Models.EF
 {
     public partial class IptColorP
     {
-        public decimal Id { get; set; }
+        public long Id { get; set; }
         public string Description { get; set; } = null!;
-        public decimal BasicColorId { get; set; }
+        public long BasicColorId { get; set; }
         public string Abbreviation { get; set; } = null!;
         public char DoNotUseInd { get; set; }
