+++ FDB.Apollo.IPT/FDB.Apollo.IPT/Models/Color.cs
@@ -3,9 +3,9 @@
     public class Color
     {
         public PublishAudit? Audit { get; set; }
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string Description { get; set; } = null!;
-        public int BasicColorID { get; set; }
+        public long BasicColorID { get; set; }
         public string Abbreviation { get; set; } = null!;
         public bool DoNotUseInd { get; set; }
         public BasicColor? BasicColor { get; set; }
