+++ sql/ddl/IPT_BASIC_COLOR_P.sql
@@ -5,7 +5,7 @@
 -- DROP TABLE public.ipt_basic_color_p;
 
 CREATE TABLE public.ipt_basic_color_p (
-	id numeric(8) NOT NULL,
+	id bigint NOT NULL,
 	description varchar(30) NOT NULL,
 	abbreviation varchar(7) NOT NULL,
 	short_abbreviation varchar(4) NOT NULL,
