+++ sql/ddl/IPT_COLOR_W.sql
@@ -5,9 +5,9 @@
 -- DROP TABLE public.ipt_color_w;
 
 CREATE TABLE public.ipt_color_w (
-	id numeric(8) NOT NULL,
+	id bigint NOT NULL,
 	description varchar(30) NOT NULL,
-	basic_color_id numeric(8) NOT NULL,
+	basic_color_id bigint NOT NULL,
 	abbreviation varchar(15) NOT NULL,
 	do_not_use_ind bpchar(1) NOT NULL,
 	CONSTRAINT ipt_color_w_ck1 CHECK ((do_not_use_ind = ANY (ARRAY['0'::bpchar, '1'::bpchar]))),
